function/requirements.txt
Flask
pandas
sklearn
google-cloud-storage